test_api/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
autopep8==2.0.0
channels==4.0.0
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-filter==2.4.0
django-graphql-jwt==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.13.1
graphene==3.3
graphene-django==3.2.0
graphene-pagination==1.0.1
graphql-core==3.2.3
graphql-relay==3.2.0
numpy==1.24.2
pandas==2.0.0
platformdirs==4.1.0
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.10.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2021.1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.25.0
